pyproject.toml
[tool.poetry]
name = "5-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "5_test"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
djangorestframework = "^3.15.0"
djoser = "^2.2.2"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
phonenumbers = "^8.13.32"
django-phonenumber-field = "^7.3.0"
geopy = "^2.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
